Fix bounds and make Alignment inner class
