+++ src/main/java/dev/emi/emi/api/widget/TextWidget.java
@@ -5,6 +5,7 @@ import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.text.OrderedText;
 
 public class TextWidget extends Widget {
+	private static final int HEIGHT = 10;
 	private static final MinecraftClient CLIENT = MinecraftClient.getInstance();
 	private final OrderedText text;
 	private final int x, y;
@@ -33,14 +34,17 @@ public class TextWidget extends Widget {
 
 	@Override
 	public Bounds getBounds() {
-		return new Bounds(x, y, CLIENT.textRenderer.getWidth(text), 10);
+		int width = CLIENT.textRenderer.getWidth(text);
+		int xOff = horizontalAlignment.offset(width);
+		int yOff = verticalAlignment.offset(HEIGHT);
+		return new Bounds(x + xOff, y + yOff, width, HEIGHT);
 	}
 
 	@Override
 	public void render(MatrixStack matrices, int mouseX, int mouseY, float delta) {
 		matrices.push();
 		int xOff = horizontalAlignment.offset(CLIENT.textRenderer.getWidth(text));
-		int yOff = verticalAlignment.offset(CLIENT.textRenderer.fontHeight);
+		int yOff = verticalAlignment.offset(HEIGHT);
 		matrices.translate(xOff, yOff, 300);
 		if (shadow) {
 			CLIENT.textRenderer.drawWithShadow(matrices, text, x, y, color);
@@ -48,5 +52,17 @@ public class TextWidget extends Widget {
 			CLIENT.textRenderer.draw(matrices, text, x, y, color);
 		}
 		matrices.pop();
+	}
+
+	public enum Alignment {
+		START, CENTER, END;
+
+		public int offset(int length) {
+			return switch (this) {
+				case START -> 0;
+				case CENTER -> -(length / 2);
+				case END -> -length;
+			};
+		}
 	}
 }
